requirements.txt
django
psycopg2-binary
flask